refactor(looker): build dashboard specs in a utility function (#21833)
## Summary & Motivation
Addresses
https://github.com/dagster-io/dagster/pull/21687#discussion_r1592844429.

## How I Tested These Changes
pytest
